+++ packages/react-ui/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @phantom/react-ui
 
+## 2.0.0-beta.10
+
+### Patch Changes
+
+- 53f1db0: Support auth user id
+- Updated dependencies [53f1db0]
+  - @phantom/browser-sdk@1.0.0-beta.15
+  - @phantom/react-sdk@1.0.0-beta.15
+  - @phantom/client@1.0.0-beta.15
+
 ## 2.0.0-beta.9
 
 ### Patch Changes